Applied Pull Up Method Refactoring
