+++ src/com/inventory/dao/BuildTableModel.java
@@ -1,0 +1,36 @@
+package com.inventory.dao;
+
+import javax.swing.table.DefaultTableModel;
+import java.sql.ResultSet;
+import java.sql.ResultSetMetaData;
+import java.sql.SQLException;
+import java.util.Vector;
+
+public class BuildTableModel {
+
+    /***
+     * Refactoring name: PULL UP METHOD
+     * To remove duplication of code for the method buildTableModel() in both classes UserDAO.java and SupplierDAO.java,
+     * Pull up method refactoring is performed and method is pulled from both classes and is kept in the new class BuildTableModel.java class
+     * The class BuildTableModel.java is then extended to two classes UserDAO.java and SupplierDAO.java.
+     */
+    public DefaultTableModel buildTableModel(ResultSet rs) throws SQLException {
+        ResultSetMetaData metaData = rs.getMetaData(); //resultset ko metadata
+        Vector<String> columnNames = new Vector<String>();
+        int columnCount = metaData.getColumnCount();
+
+        for (int column = 1; column <= columnCount; column++) {
+            columnNames.add(metaData.getColumnName(column));
+        }
+
+        Vector<Vector<Object>> data = new Vector<Vector<Object>>();
+        while (rs.next()) {
+            Vector<Object> vector = new Vector<Object>();
+            for (int columnIndex = 1; columnIndex <= columnCount; columnIndex++) {
+                vector.add(rs.getObject(columnIndex));
+            }
+            data.add(vector);
+        }
+        return new DefaultTableModel(data, columnNames);
+    }//end of method DefaultTableModel
+}
