+++ src/com/inventory/dao/SupplierDAO.java
@@ -11,18 +11,17 @@ import com.inventory.dto.SupplierDTO;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.Vector;
 import javax.swing.JOptionPane;
-import javax.swing.table.DefaultTableModel;
 
-/**
- *
- * @author ADMIN
+/***
+ * Refactoring name: PULL UP METHOD
+ * To remove duplication of code for the method buildTableModel() in both classes UserDAO.java and SupplierDAO.java,
+ * Pull up method refactoring is performed and method is pulled from both classes and is kept in the new class BuildTableModel.java class
+ * The class BuildTableModel.java is then extended to two classes UserDAO.java and SupplierDAO.java.
  */
-public class SupplierDAO {
+public class SupplierDAO extends BuildTableModel{
     Connection con = null;
     PreparedStatement pstmt = null;
     Statement stmt = null;
@@ -152,24 +151,4 @@ public class SupplierDAO {
         return rs;
     }
 
-    //start of method DefaultTableModel
-    public DefaultTableModel buildTableModel(ResultSet rs) throws SQLException {
-        ResultSetMetaData metaData = rs.getMetaData(); //resultset ko metadata
-        Vector<String> columnNames = new Vector<String>();
-        int columnCount = metaData.getColumnCount();
-
-        for (int column = 1; column <= columnCount; column++) {
-            columnNames.add(metaData.getColumnName(column));
-        }
-
-        Vector<Vector<Object>> data = new Vector<Vector<Object>>();
-        while (rs.next()) {
-            Vector<Object> vector = new Vector<Object>();
-            for (int columnIndex = 1; columnIndex <= columnCount; columnIndex++) {
-                vector.add(rs.getObject(columnIndex));
-            }
-            data.add(vector);
-        }
-        return new DefaultTableModel(data, columnNames);
-    }//end of method DefaultTableModel
 }
